TyphoonInitializer refactored to new API

+++ Source/Definition/Injections/TyphoonAbstractInjection.m
@@ -18,7 +18,7 @@
     if (self.type == TyphoonInjectionTypeUndefinied) {
         return [NSString stringWithFormat:@"<%@: %p, type=Undifined>",[self class], self];
     } else if (self.type == TyphoonInjectionTypeParameter) {
-        return [NSString stringWithFormat:@"<%@: %p, index=%d>",[self class], self, self.parameterIndex];
+        return [NSString stringWithFormat:@"<%@: %p, index=%d>",[self class], self, (int)self.parameterIndex];
     } else {
         return [NSString stringWithFormat:@"<%@: %p, property=%@>",[self class], self, self.propertyName];
     }
